requirements-ai.txt
# Optional AI provider dependencies.
#
# Gemini uses a local SDK and image helpers. Claude and Grok integrations use
# direct HTTPS APIs and do not require additional client libraries here.
google-genai==1.66.0
